feat(router) add redirectToFirstAllowNode to routes, witch will be transfer to first canAccess children route (#29)
Added redirectToFirstAllowNode to route config which will be redirect to first allowed child nodes.
If no one node is allowed, it will be redirected to NOT_FOUND or root

When it can help you:

Your role model assumes a different set of routers for different roles.
You don't know where to redirect the user because the config is built from different modules